requirements.txt
customtkinter
requests
keyboard
pywin32
pystray
pillow